chore: day 6

+++ src/main/kotlin/com/denormans/adventofcode/utils/CollectionUtils.kt
@@ -1,7 +1,7 @@
 package com.denormans.adventofcode.utils
 
-fun <T> Collection<T>.withCount(): MutableMap<T, Int> {
-  val pointsWithCount = mutableMapOf<T, Int>()
+fun <T> Collection<T>.withCount(): MutableMap<T, Long> {
+  val pointsWithCount = mutableMapOf<T, Long>()
   forEach { pointsWithCount[it] = pointsWithCount.getOrDefault(it, 0) + 1 }
   return pointsWithCount
 }
